pyproject.toml
[tool.poetry]
name = "stock_market_strategies"
version = "0.1.0"
description = ""
authors = ["Akash <akash.iitr17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
matplotlib = "^3.9.2"
ipykernel = "^6.29.5"
pandas = "^2.2.2"
pyyaml = "^6.0.2"
tqdm = "^4.66.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
